[ADD] stock_deferred_assign: Add new module TT31143
