+++ stock_deferred_assign/models/__init__.py
@@ -2,4 +2,5 @@
 from . import sale
 from . import stock_move
 from . import stock_picking
+from . import stock_picking_type
 from . import stock_rule
