+++ stock_deferred_assign/models/stock_picking.py
@@ -6,27 +6,6 @@ from datetime import timedelta
 from odoo import fields, models
 
 
-class PickingType(models.Model):
-    _inherit = "stock.picking.type"
-
-    reservation_method = fields.Selection(
-        [
-            ("at_confirm", "At Confirmation"),
-            ("manual", "Manually"),
-            ("by_date", "Before scheduled date"),
-        ],
-        "Reservation Method",
-        required=True,
-        default="at_confirm",
-        help="How products in transfers of this operation type should be reserved.",
-    )
-    reservation_days_before = fields.Integer(
-        "Days",
-        help="Maximum number of days before scheduled date that products "
-        "should be reserved.",
-    )
-
-
 class StockPicking(models.Model):
     _inherit = "stock.picking"
