+++ stock_deferred_assign/models/stock_picking_type.py
@@ -1,0 +1,26 @@
+# Copyright 2021 Tecnativa - Carlos Dauden
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
+
+from odoo import fields, models
+
+
+class PickingType(models.Model):
+    _inherit = "stock.picking.type"
+
+    reservation_method = fields.Selection(
+        [
+            ("at_confirm", "At Confirmation"),
+            ("manual", "Manually"),
+            ("by_date", "Before scheduled date"),
+        ],
+        "Reservation Method",
+        required=True,
+        default="at_confirm",
+        help="How products in transfers of this operation type should be reserved.",
+    )
+    reservation_days_before = fields.Integer(
+        "Days",
+        help="Maximum number of days before scheduled date that products "
+        "should be reserved.",
+    )
