+++ stock_deferred_assign/readme/CONFIGURATION.rst
@@ -1,0 +1,2 @@
+#. Go to `Inventory > Configuration > Operations Types` and edit/create one
+#. Set `Reservation Method` option.
