+++ stock_deferred_assign/readme/DESCRIPTION.rst
@@ -1,1 +1,2 @@
-This module allows
+This module allows you to configure at the picking type level the moment in which
+the stock reservation is done.
